lezen en wegschrijven van tekstbestanden is voltooid

+++ Opdracht1/src/DatumV1/Datum.java
@@ -196,9 +196,9 @@ public class Datum implements Comparable<Datum>
 	public int verschilInMaanden(Datum d)
 	{			
 		if (this.compareTo(d) < 0)
-			return -(int)this.compareTo(d)/(365/12);
+			return -this.compareTo(d)/(365/12);
 		else			
-			return (int)this.compareTo(d)/(365/12);
+			return this.compareTo(d)/(365/12);
 	}
 	
 	public int verschilInDagen(Datum d)
